+++ Tickette.Domain/Entities/Payment.cs
@@ -2,17 +2,17 @@
 
 public class Payment
 {
-    public long Amount { get; set; }
+    public decimal Amount { get; set; }
 
     public string Currency { get; set; }
 
-    private Payment(long amount, string currency)
+    private Payment(decimal amount, string currency)
     {
         Amount = amount;
         Currency = currency;
     }
 
-    public static Payment Create(long amount, string currency = "usd")
+    public static Payment Create(decimal amount, string currency = "usd")
     {
         return new Payment(amount, currency);
     }
